Fix gestión de errores

+++ server-springboot-micros/server-springboot-author/tutorialauthor/src/main/java/com/ludoteca/tutorialauthor/author/AuthorServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -42,6 +43,13 @@ public class AuthorServiceImpl implements AuthorService {
      */
     @Override
     public void save(Long id, AuthorDto data) {
+        Optional<Author> existingAuthor = this.authorRepository.findByName(data.getName());
+
+        if (existingAuthor.isPresent()) {
+            if (id == null || !existingAuthor.get().getId().equals(id)) {
+                throw new org.springframework.web.server.ResponseStatusException(org.springframework.http.HttpStatus.CONFLICT, "El autor con el nombre '" + data.getName() + "' ya existe.");
+            }
+        }
 
         Author author;
 
@@ -60,10 +68,17 @@ public class AuthorServiceImpl implements AuthorService {
      * {@inheritDoc}
      */
     @Override
-    public void delete(Long id) throws Exception {
+    public void delete(Long id) {
 
         if (this.get(id) == null) {
-            throw new Exception("Not exists");
+            throw new org.springframework.web.server.ResponseStatusException(org.springframework.http.HttpStatus.NOT_FOUND, "Not exists");
+        }
+
+        String url = "http://localhost:8080/game/exists-author/" + id;
+        Boolean hasGames = new org.springframework.web.client.RestTemplate().getForObject(url, Boolean.class);
+
+        if (Boolean.TRUE.equals(hasGames)) {
+            throw new org.springframework.web.server.ResponseStatusException(org.springframework.http.HttpStatus.BAD_REQUEST, "El autor tiene juegos asociados");
         }
 
         this.authorRepository.deleteById(id);
